backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.13
alembic==1.14.0
pydantic-settings==2.6.1
python-dotenv==1.0.1
redis==5.2.0
PyJWT==2.10.1
qdrant-client==1.12.1
openai==1.54.3
anthropic==0.39.0
APScheduler==3.10.4
python-dateutil==2.9.0.post0

